refactor: :recycle: refactor view structure componets
