+++ src/app/modules/feedback/views/feedback-detail/feedback-detail.component.scss
