--- src/app/modules/home/layout/home.component.spec.ts
@@ -1,27 +1,0 @@
-import { ComponentFixture, TestBed } from '@angular/core/testing';
-import { RouterTestingModule } from '@angular/router/testing';
-import { FeedbackService } from '@core/http/feedback.service';
-import { HomeComponent } from './home.component';
-
-describe('SuggestionsLayoutComponent', () => {
-    let component: HomeComponent;
-    let fixture: ComponentFixture<HomeComponent>;
-
-    beforeEach(async () => {
-        await TestBed.configureTestingModule({
-            declarations: [HomeComponent],
-            imports: [RouterTestingModule],
-            providers: [FeedbackService],
-        }).compileComponents();
-    });
-
-    beforeEach(() => {
-        fixture = TestBed.createComponent(HomeComponent);
-        component = fixture.componentInstance;
-        fixture.detectChanges();
-    });
-
-    it.skip('should create', () => {
-        expect(component).toBeTruthy();
-    });
-});
