+++ src/app/modules/home/views/home-view/home-view.component.spec.ts
@@ -1,0 +1,24 @@
+import { ComponentFixture, TestBed } from '@angular/core/testing';
+
+import { HomeViewComponent } from './home-view.component';
+
+describe('HomeViewComponent', () => {
+    let component: HomeViewComponent;
+    let fixture: ComponentFixture<HomeViewComponent>;
+
+    beforeEach(async () => {
+        await TestBed.configureTestingModule({
+            declarations: [HomeViewComponent],
+        }).compileComponents();
+    });
+
+    beforeEach(() => {
+        fixture = TestBed.createComponent(HomeViewComponent);
+        component = fixture.componentInstance;
+        fixture.detectChanges();
+    });
+
+    it('should create', () => {
+        expect(component).toBeTruthy();
+    });
+});
