+++ src/app/modules/home/views/home-view/home-view.component.ts
@@ -2,16 +2,16 @@ import { Component, OnInit } from '@angular/core';
 import { Router } from '@angular/router';
 import { FeedbackService } from '@core/http/feedback.service';
 import { Feedback } from '@core/models/feedback.model';
-import { map, Observable } from 'rxjs';
+import { Observable, map } from 'rxjs';
 
 @Component({
-    selector: 'app-home',
-    templateUrl: './home.component.html',
-    styleUrls: ['./home.component.scss'],
+    selector: 'app-home-view',
+    templateUrl: './home-view.component.html',
+    styleUrls: ['./home-view.component.scss'],
 })
-export class HomeComponent implements OnInit {
+export class HomeViewComponent implements OnInit {
     feedbacks$!: Observable<Feedback[]>;
-    category:string = 'all';
+    category: string = 'all';
 
     constructor(private feedbacks: FeedbackService, private router: Router) {
         this.feedbacks$ = feedbacks.getFeedbackList();
@@ -27,7 +27,7 @@ export class HomeComponent implements OnInit {
     }
 
     feebacksByStatus(status: string) {
-        return this.feedbacks$.pipe(map(feedbacks => feedbacks.filter(feedback => feedback.status === status)));
+        return this.feedbacks$.pipe(map((feedbacks) => feedbacks.filter((feedback) => feedback.status === status)));
     }
 
     private getFeedbacksFiltered(category: string): Observable<Feedback[]> {
