+++ src/app/modules/roadmap/roadmap.module.ts
@@ -1,6 +1,6 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
-import { RoadmapDetailComponent } from './layouts/roadmap-detail/roadmap-detail.component';
+import { RoadmapDetailComponent } from './views/roadmap-detail/roadmap-detail.component';
 import { RouterModule } from '@angular/router';
 import { UiToolbarModule } from '@components/toolbar/ui-toolbar.module';
 import { UiButtonModule } from '@components/button/ui-button.module';
@@ -14,30 +14,26 @@ import { UiBadgeModule } from '@components/badge/ui-badge.module';
 import { UiChipsModule } from '@components/chips/ui-chips.module';
 
 const routes = [
-  {
-    path: ':category',
-    component: RoadmapDetailComponent,
-    pathMatch: 'full'
-  }
-]
+    {
+        path: ':category',
+        component: RoadmapDetailComponent,
+        pathMatch: 'full',
+    },
+];
 
 @NgModule({
-  declarations: [
-    RoadmapDetailComponent,
-    KanbanColumnComponent,
-    KanbanBoardComponent,
-  ],
-  imports: [
-    CommonModule,
-    RouterModule.forChild(routes),
-    TranslocoModule,
-    UiToolbarModule,
-    UiButtonModule,
-    UiIconModule,
-    KanbanCardModule,
-    UiCounterModule,
-    UiBadgeModule,
-    UiChipsModule
-  ]
+    declarations: [RoadmapDetailComponent, KanbanColumnComponent, KanbanBoardComponent],
+    imports: [
+        CommonModule,
+        RouterModule.forChild(routes),
+        TranslocoModule,
+        UiToolbarModule,
+        UiButtonModule,
+        UiIconModule,
+        KanbanCardModule,
+        UiCounterModule,
+        UiBadgeModule,
+        UiChipsModule,
+    ],
 })
-export class RoadmapModule { }
+export class RoadmapModule {}
